did taser UI implementation

+++ GameDesign/Assets/Scripts/MonoBehaviours/SpawnerManagerMonoBehaviour.cs
@@ -27,10 +27,10 @@ public class SpawnerManagerMonoBehaviour : MonoBehaviour
     void Start()
     {
         leverManagerSingleton = LevelSettingManager.Instance;
-        if (timerManager == null)
+        if(timerManager == null)
         {
             timerManager = GetComponent<TimerManagerMonoBehaviour>();
-            if (timerManager == null)
+            if(timerManager == null)
             {
                 Debug.LogError("Couldn't find timer manager");
             }
@@ -56,22 +56,22 @@ public class SpawnerManagerMonoBehaviour : MonoBehaviour
             {
                 if (spawnRates[t].Evaluate(timePercentage) >= p)
                 {
-                    toSpawnUnmasked[UnityEngine.Random.Range(0, levelSpawners.Length)].Add(t, 1);
+                    toSpawnUnmasked[(int)UnityEngine.Random.Range(0f, levelSpawners.Length)].Add(t, 1);
                 }
             }
             for (int i = 0; i < toSpawnUnmasked.Length; i++)
             {
                 Dictionary<CustomerTypes, int> toSpawn = toSpawnUnmasked[i];
-
+                
                 p = UnityEngine.Random.Range(0f, 1.01f);
                 int spawnMasked = maskedSpawnRate.Evaluate(timePercentage) >= p ? 1 : 0;
-
+                
                 var (spawnRequestWM, spawnRequestUnmasked) = leverManagerSingleton.RequestSpawn(spawnMasked, toSpawn);
                 levelSpawners[i].Spawn(spawnRequestWM, spawnRequestUnmasked);
             }
             currentPullRate = currentTime + pullRate;
         }
-
+        
 
     }
 }
